style(jobs): rename `CronJob` to `BaseCronJob`
